Updated Component and variable names for consistency.
Updated EventMarker to StaticMarker.
Updated Locationmarker to DynamicMarker.
Updated 'location' and 'position' to 'coordinates'.
Position is now a reserved term for the Marker component prop.
Location is (for) now a reserved term for communicating with users.
Coordinates will be the default term to refer to [lat,lng] data.
Updated LocationContext.ts to TemporaryCoordinatesContext.ts, since this context refers to the DynamicMarker coordinates.